requirements.txt
psycopg2-binary==2.9.11
python-docx==1.2.0
chromadb>=0.4.0
sentence-transformers>=2.2.0
langchain>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.1.0
langchain-core>=0.1.0
httpx>=0.25.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
